immobilier-backend/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.6.0
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cssselect==1.1.0
Django==4.1
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.9.0
django-nine==0.2.5
djangorestframework==3.13.1
djongo==1.3.6
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.1
mypy-extensions==0.4.3
parse==1.19.0
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pyee==8.2.2
pymongo==3.12.1
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.2
requests==2.28.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
w3lib==2.0.1
websockets==10.3
zipp==3.8.1
